rebased main, added allow-schema-migraion flag, added indexes creation, removal tests, refactored enum syncing
