+++ packages/node-core/src/utils/index.ts
@@ -13,3 +13,4 @@ export * from './blockSizeBuffer';
 export * from './configure';
 export * from './reindex';
 export * from './decorators';
+export * from './sequelizeUtil';
